DEV-222: Changed type of TxOut.Value and util.Amount to uin64 (#108)
* [DEV-222] Changed type of TxOut.Value and util.Amount to uin64

* [DEV-222] Replaced math.MaxUint64 with 0

* [DEV-222] Fixed comment to reflect uint64 instead of int64

* [DEV-222] Fixed overflow comment

+++ blockdag/dagio.go
@@ -555,7 +555,7 @@ func deserializeUTXOEntry(serialized []byte) (*UTXOEntry, error) {
 	}
 
 	entry := &UTXOEntry{
-		amount:      int64(amount),
+		amount:      amount,
 		pkScript:    pkScript,
 		blockHeight: blockHeight,
 		packedFlags: 0,
